12_OpenAI_Agent_SDK/13_LiteLLM_OpenRouter/pyproject.toml
[project]
name = "13-litellm-openrouter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents[litellm]>=0.1.0",
    "python-dotenv>=1.1.1",
]
